refactor: Update html driver test with the updated tests
Refactor html tests to separate React markups and tests, so that the tests can run well in Vite.
