+++ packages/component-driver-html-test/src/directory.tsx
@@ -1,13 +1,19 @@
+import { JSX } from 'react';
+
 import { ExampleList } from './components/ExampleList';
-import {
-  checkboxExamples,
-  focusEventExamples,
-  formExamples,
-  mouseEventExamples,
-  radioButtonGroupExamples,
-  selectExamples,
-  textInputExamples,
-} from './examples';
+import { checkboxGroupUIExample } from './examples/checkbox/CheckboxGroup.examples';
+import { singleCheckboxUIExample } from './examples/checkbox/SingleCheckbox.examples';
+import { focusEventUIExample } from './examples/focusEvent/Focus.examples';
+import { linkedElementUIExample } from './examples/form/LinkedElement.examples';
+import { controlledTextInputUIExample } from './examples/input/Controlled.examples';
+import { uncontrolledTextInputUIExample } from './examples/input/Uncontrolled.examples';
+import { clickLocationMouseEventUIExample } from './examples/mouseEvent/ClickLocation.examples';
+import { hoverMouseEventUIExample } from './examples/mouseEvent/Hover.examples';
+import { mouseLocationMouseEventUIExample } from './examples/mouseEvent/MouseLocation.examples';
+import { mouseOverMouseEventUIExample } from './examples/mouseEvent/MouseOver.examples';
+import { uncontrolledRadioButtonGroupUIExample } from './examples/radioButtonGroup/Uncontrolled.examples';
+import { multipleSelectUIExample } from './examples/select/MultipleSelect.examples';
+import { singleSelectUIExample } from './examples/select/SingleSelect.examples';
 
 interface IToc {
   label: string;
@@ -19,36 +25,45 @@ export const tocs: IToc[] = [
   {
     label: 'Form',
     path: '/form',
-    ui: <ExampleList examples={formExamples} />,
+    ui: <ExampleList examples={[linkedElementUIExample]} />,
   },
   {
     label: 'Text Input',
     path: '/input',
-    ui: <ExampleList examples={textInputExamples} />,
+    ui: <ExampleList examples={[uncontrolledTextInputUIExample, controlledTextInputUIExample]} />,
   },
   {
     label: 'Radio Buttons',
     path: '/radio-buttons',
-    ui: <ExampleList examples={radioButtonGroupExamples} />,
+    ui: <ExampleList examples={[uncontrolledRadioButtonGroupUIExample]} />,
   },
   {
     label: 'Checkbox',
     path: '/checkbox',
-    ui: <ExampleList examples={checkboxExamples} />,
+    ui: <ExampleList examples={[singleCheckboxUIExample, checkboxGroupUIExample]} />,
   },
   {
     label: 'Select',
     path: '/select',
-    ui: <ExampleList examples={selectExamples} />,
+    ui: <ExampleList examples={[singleSelectUIExample, multipleSelectUIExample]} />,
   },
   {
     label: 'Focus Event',
     path: '/focus-event',
-    ui: <ExampleList examples={focusEventExamples} />,
+    ui: <ExampleList examples={[focusEventUIExample]} />,
   },
   {
     label: 'Mouse Event',
     path: '/mouse-event',
-    ui: <ExampleList examples={mouseEventExamples} />,
+    ui: (
+      <ExampleList
+        examples={[
+          hoverMouseEventUIExample,
+          clickLocationMouseEventUIExample,
+          mouseLocationMouseEventUIExample,
+          mouseOverMouseEventUIExample,
+        ]}
+      />
+    ),
   },
 ];
