+++ packages/component-driver-html-test/src/examples/focusEvent/Focus.examples.tsx
@@ -1,8 +1,6 @@
-import React, { useCallback } from 'react';
+import React, { useCallback, JSX } from 'react';
 
-import { HTMLElementDriver, HTMLTextInputDriver } from '@atomic-testing/component-driver-html';
-import { IExampleUnit, ScenePart, TestEngine, byDataTestId } from '@atomic-testing/core';
-import { TestSuiteInfo } from '@atomic-testing/test-runner';
+import { IExampleUIUnit } from '@atomic-testing/core';
 
 export const FocusEventExample = () => {
   const [showDetail, setShowDetail] = React.useState<string>('');
@@ -29,63 +27,7 @@ export const FocusEventExample = () => {
   );
 };
 
-export const focusEventExampleScenePart = {
-  target: {
-    locator: byDataTestId('focus-target'),
-    driver: HTMLTextInputDriver,
-  },
-  blurAid: {
-    locator: byDataTestId('blur-aid'),
-    driver: HTMLTextInputDriver,
-  },
-  detail: {
-    locator: byDataTestId('focus-detail'),
-    driver: HTMLElementDriver,
-  },
-} satisfies ScenePart;
-
-export const focusEventExample: IExampleUnit<typeof focusEventExampleScenePart, JSX.Element> = {
+export const focusEventUIExample: IExampleUIUnit<JSX.Element> = {
   title: 'Focus event',
-  scene: focusEventExampleScenePart,
   ui: <FocusEventExample />,
-};
-
-export const focusEventExampleTestSuite: TestSuiteInfo<typeof focusEventExample.scene> = {
-  title: 'Focus Event: Focus',
-  url: '/focus-event',
-  tests: (getTestEngine, { describe, test, beforeEach, afterEach, assertEqual }) => {
-    describe(`${focusEventExample.title}`, () => {
-      let testEngine: TestEngine<typeof focusEventExample.scene>;
-
-      // Use the following beforeEach to work around the issue of Playwright's page being undefined
-      // @ts-ignore
-      beforeEach(function ({ page }) {
-        // @ts-ignore
-        testEngine = getTestEngine(focusEventExample.scene, { page });
-        if (typeof arguments[0] === 'function') {
-          arguments[0]();
-        }
-      });
-
-      afterEach(async () => {
-        await testEngine.cleanUp();
-      });
-
-      test(`Initially detail is empty`, async () => {
-        assertEqual(await testEngine.parts.detail.getText(), '');
-      });
-
-      test(`Detail is "focus" when element is focused`, async () => {
-        await testEngine.parts.target.focus();
-        assertEqual(await testEngine.parts.detail.getText(), 'focus');
-      });
-
-      test(`Detail is "blur" when element is blurred`, async () => {
-        await testEngine.parts.target.focus();
-        // Focus on another element to blur the target
-        await testEngine.parts.blurAid.focus();
-        assertEqual(await testEngine.parts.detail.getText(), 'blur');
-      });
-    });
-  },
 };
