+++ packages/component-driver-html-test/src/examples/focusEvent/index.ts
@@ -1,6 +1,8 @@
+import { JSX } from 'react';
+
 import { IExampleUnit, ScenePart } from '@atomic-testing/core';
 
-import { focusEventExample, focusEventExampleTestSuite } from './Focus.examples';
+import { focusEventExample, focusEventExampleTestSuite } from './Focus.suite';
 
 export { focusEventExample, focusEventExampleTestSuite };
