+++ packages/component-driver-html-test/src/examples/mouseEvent/ClickLocation.suite.ts
@@ -1,0 +1,67 @@
+import { JSX } from 'react';
+
+import { HTMLButtonDriver, HTMLElementDriver } from '@atomic-testing/component-driver-html';
+import { IExampleUnit, ScenePart, TestEngine, byDataTestId } from '@atomic-testing/core';
+import { TestSuiteInfo } from '@atomic-testing/test-runner';
+
+import { clickLocationMouseEventUIExample } from './ClickLocation.examples';
+
+export const clickLocationMouseEventExampleScenePart = {
+  target: {
+    locator: byDataTestId('click-target'),
+    driver: HTMLButtonDriver,
+  },
+  xDisplay: {
+    locator: byDataTestId('position-x'),
+    driver: HTMLElementDriver,
+  },
+  yDisplay: {
+    locator: byDataTestId('position-y'),
+    driver: HTMLElementDriver,
+  },
+} satisfies ScenePart;
+
+export const clickLocationMouseEventExample: IExampleUnit<typeof clickLocationMouseEventExampleScenePart, JSX.Element> =
+  {
+    ...clickLocationMouseEventUIExample,
+    scene: clickLocationMouseEventExampleScenePart,
+  };
+
+export const clickLocationMouseEventExampleTestSuite: TestSuiteInfo<typeof clickLocationMouseEventExample.scene> = {
+  title: 'Mouse event: Click',
+  url: '/mouse-event',
+  tests: (getTestEngine, { describe, test, beforeEach, afterEach, assertEqual }) => {
+    describe(`${clickLocationMouseEventExample.title}`, () => {
+      let testEngine: TestEngine<typeof clickLocationMouseEventExample.scene>;
+
+      // Use the following beforeEach to work around the issue of Playwright's page being undefined
+      // @ts-ignore
+      beforeEach(function ({ page }) {
+        // @ts-ignore
+        testEngine = getTestEngine(clickLocationMouseEventExample.scene, { page });
+        if (typeof arguments[0] === 'function') {
+          arguments[0]();
+        }
+      });
+
+      afterEach(async () => {
+        await testEngine.cleanUp();
+      });
+
+      test('Click on somewhere in the target should display the correct coordinates', async () => {
+        await testEngine.parts.target.click({
+          position: {
+            x: 10,
+            y: 5,
+          },
+        });
+        const xDisplay = await testEngine.parts.xDisplay.getText();
+        const yDisplay = await testEngine.parts.yDisplay.getText();
+
+        // The coordinates are rounded because e2e tests are not pixel perfect
+        assertEqual(Math.round(parseFloat(xDisplay ?? '')), 10);
+        assertEqual(Math.round(parseFloat(yDisplay ?? '')), 5);
+      });
+    });
+  },
+};
