+++ packages/component-driver-html-test/src/examples/mouseEvent/index.ts
@@ -1,9 +1,11 @@
+import { JSX } from 'react';
+
 import { IExampleUnit, ScenePart } from '@atomic-testing/core';
 
-import { clickLocationMouseEventExample, clickLocationMouseEventExampleTestSuite } from './ClickLocation.examples';
-import { hoverMouseEventExample, hoverMouseEventExampleTestSuite } from './Hover.examples';
-import { mouseLocationMouseEventExample, mouseLocationMouseEventExampleTestSuite } from './MouseLocation.examples';
-import { mouseOverMouseEventExample, mouseOverMouseEventExampleTestSuite } from './MouseOver.examples';
+import { clickLocationMouseEventExample, clickLocationMouseEventExampleTestSuite } from './ClickLocation.suite';
+import { hoverMouseEventExample, hoverMouseEventExampleTestSuite } from './Hover.suite';
+import { mouseLocationMouseEventExample, mouseLocationMouseEventExampleTestSuite } from './MouseLocation.suite';
+import { mouseOverMouseEventExample, mouseOverMouseEventExampleTestSuite } from './MouseOver.suite';
 
 export {
   clickLocationMouseEventExample,
