+++ packages/component-driver-html-test/src/examples/select/index.ts
@@ -1,13 +1,12 @@
+import { JSX } from 'react';
+
 import { IExampleUnit, ScenePart } from '@atomic-testing/core';
 
-import { multipleSelectExample as multipleSelect } from './MultipleSelect.examples';
-import { singleSelectExample as singleSelect } from './SingleSelect.examples';
+import { multipleSelectExample, multipleSelectTestSuite } from './MultipleSelect.suite';
+import { singleSelectExample, singleSelectTestSuite } from './SingleSelect.suite';
 
-export { multipleSelectTestSuite } from './MultipleSelect.examples';
-export { singleSelectTestSuite } from './SingleSelect.examples';
-
-export const multipleSelectExample = multipleSelect;
-export const singleSelectExample = singleSelect;
+export { multipleSelectTestSuite, singleSelectTestSuite };
+export { multipleSelectExample, singleSelectExample };
 
 export const selectExamples = [singleSelectExample, multipleSelectExample] satisfies IExampleUnit<
   ScenePart,
